security-toolkit/tool-11-port-scanner/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-nmap==0.7.1
requests==2.31.0
reportlab==4.0.7
